Fix how we set initiative needs work
